Progressing towards es6 module.
