+++ lib/HttpSignatureError.js
@@ -3,7 +3,7 @@
  */
 'use strict';
 
-module.exports = class HTTPSignatureError extends Error {
+export default class HTTPSignatureError extends Error {
   constructor(message, name) {
     super(message);
     this.name = name;
