+++ lib/env.js
@@ -4,12 +4,7 @@
 'use strict';
 
 // determine if using node.js or browser
-const nodejs = (
+export const nodejs = (
   typeof process !== 'undefined' && process.versions && process.versions.node);
-const browser = !nodejs &&
+export const browser = !nodejs &&
   (typeof window !== 'undefined' || typeof self !== 'undefined');
-
-module.exports = {
-  nodejs,
-  browser
-};
